Added schedule previews to medication list
