+++ app/src/main/res/layout/schedule_record.xml
@@ -8,7 +8,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constraintLayout"
         android:layout_width="fill_parent"
-        android:layout_height="32dp"
+        android:layout_height="28dp"
         android:layout_marginBottom="8dp"
         app:layout_constraintBottom_toBottomOf="parent"
         tools:layout_editor_absoluteX="64dp">
